Update yarn.lock with latest package versions
This commit updates the versions of several packages in the yarn.lock file. For example, the '@babel/runtime' options are upgraded from "7.23.2" to "7.23.6". This process is crucial to ensure the application stays updated with the latest available packages, which often come with essential patches, improvements, and additional functionalities. All the changes align with the package version updates and no new package is added or removed.